requirements.txt
# Database
psycopg2-binary

# Web Scraping
beautifulsoup4
requests

# LangChain & LLMs
langchain
langchain-openai
langchain-groq
langchain-core
langgraph

# API Framework
fastapi
uvicorn[standard]
pydantic

# Frontend
streamlit

# Utilities
python-dotenv